+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Geolocation/GeolocationLayers.cs
@@ -26,8 +26,8 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Geolocation
         public GeolocationLayers(string mapId, IEnumerable<MapEvent>? positionEvents = null, IEnumerable<MapEvent>? accuracyEvents = null)
         {
             MapId = mapId;
-            positionLayer = new SymbolLayer() { DataSource = dataSourceDef };
-            accuracyLayer = new PolygonLayer() { DataSource = dataSourceDef };
+            positionLayer = new SymbolLayer() { Source = dataSourceDef };
+            accuracyLayer = new PolygonLayer() { Source = dataSourceDef };
 
             //Render point or MultiPoints in this layer.
             positionLayer.Options ??= new();