+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Layers/PolygonExtrusionLayer.cs
@@ -6,7 +6,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
     /// <summary>
     /// <inheritdoc cref="LayerType.PolygonExtrusion"/>
     /// </summary>
-    public class PolygonExtrusionLayer : LayerBase<PolygonExtrusionLayerOptions, DataSource>
+    public class PolygonExtrusionLayer : LayerBase<PolygonExtrusionLayerOptions>
     {
         /// <summary>
         /// <inheritdoc cref="LayerType"/>