+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/scripts/modules/Sources.ts
@@ -18,34 +18,42 @@ export class Sources {
     }
 
     static #doAdd(mapRef: MapReference, def: Source, events?: MapEvent[]): void {
+        if (Helpers.isEmptyOrNull(def.id)) {
+            Logger.logMessage(mapRef.mapId, LogLevel.Error, `Sources.add: missing Id.`, def);
+            return;
+        }
+
+        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `Sources.add: searching for source with ID '${def.id}'.`);
+        let ds = mapRef.map.sources.getById(def.id);
+
+        if (ds) {
+            Logger.logMessage(mapRef.mapId, LogLevel.Warn, `Sources.add: source with ID '${def.id}' already exists.`);
+            return;
+        }
+
         const sourceType = Helpers.switchCaseInsensitive(def.type, {
             data: () => "Data",
+            elevationtile: () => "ElevationTile",
+            vectortile: () => "VectorTile"
         }, () => "");
 
         switch (sourceType) {
             case 'Data':
                 Sources.#doAddDataSource(mapRef, def as DataSource, events);
                 break;
+            case 'ElevationTile':
+                Sources.#doAddElevationTileSource(mapRef, def as ElevationTileSource, events);
+                break;
+            case 'VectorTile':
+                Sources.#doAddVectorTileSource(mapRef, def as VectorTileSource, events);
+                break;
             default:
-                Logger.logMessage(mapRef.mapId, LogLevel.Warn, `add: unsupported source type: ${def.type}`);
+                Logger.logMessage(mapRef.mapId, LogLevel.Warn, `Sources.add: unsupported source type: ${def.type}`);
                 return;
         }
     }
 
     static #doAddDataSource(mapRef: MapReference, def: DataSource, events?: MapEvent[]): void {
-        if (Helpers.isEmptyOrNull(def.id)) {
-            Logger.logMessage(mapRef.mapId, LogLevel.Error, `addDataSource: missing Id.`, def);
-            return;
-        }
-
-        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addDataSource: searching for source with ID '${def.id}'.`);
-        let ds = mapRef.map.sources.getById(def.id);
-
-        if (ds) {
-            Logger.logMessage(mapRef.mapId, LogLevel.Warn, `addDataSource: source with ID '${def.id}' already exists.`);
-            return;
-        }
-
         //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addDataSource: creating source with ID '${def.id}'.`);
         const newDs = new atlas.source.DataSource(def.id, def.options);
 
@@ -54,6 +62,30 @@ export class Sources {
         }
 
         //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addDataSource: adding source with ID '${def.id}'.`);
+        mapRef.map.sources.add(newDs);
+    }
+
+    static #doAddElevationTileSource(mapRef: MapReference, def: ElevationTileSource, events?: MapEvent[]): void {
+        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addElevationTileSource: creating source with ID '${def.id}'.`);
+        const newDs = new atlas.source.ElevationTileSource(def.id, def.options);
+
+        if (events) {
+            Events.sources.add(mapRef, events, newDs);
+        }
+
+        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addElevationTileSource: adding source with ID '${def.id}'.`);
+        mapRef.map.sources.add(newDs);
+    }
+
+    static #doAddVectorTileSource(mapRef: MapReference, def: VectorTileSource, events?: MapEvent[]): void {
+        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addVectorTileSource: creating source with ID '${def.id}'.`);
+        const newDs = new atlas.source.VectorTileSource(def.id, def.options);
+
+        if (events) {
+            Events.sources.add(mapRef, events, newDs);
+        }
+
+        //Logger.logMessage(mapRef.mapId, LogLevel.Trace, `addVectorTileSource: adding source with ID '${def.id}'.`);
         mapRef.map.sources.add(newDs);
     }
     // #endregion
@@ -217,3 +249,10 @@ export interface DataSource extends Source {
     options?: atlas.DataSourceOptions;
 }
 
+export interface ElevationTileSource extends Source {
+    options?: atlas.ElevationTileSourceOptions;
+}
+
+export interface VectorTileSource extends Source {
+    options?: atlas.VectorTileSourceOptions;
+}